完成修改飲食帖功能

+++ DietDiary/API/FoodsData.swift
@@ -24,11 +24,11 @@ class FoodsData: BaseResponseData {
 class Food: Mappable {
     var category: String?
     var name: String?
-    var weight: Int?
-    var calories: Int?
-    var protein: Int?
-    var fat: Int?
-    var carbohydrate: Int?
+    var weight: Float?
+    var calories: Float?
+    var protein: Float?
+    var fat: Float?
+    var carbohydrate: Float?
     
     required init?(map: Map) {}
     
@@ -42,18 +42,18 @@ class Food: Mappable {
         carbohydrate <- map["carbohydrate"]
     }
     
-    var grains: Double?
-    var meats: Double?
-    var oils: Double?
-    var milk: Double?
-    var vegetables: Double?
-    var fruits: Double?
+    var grains: Float?
+    var meats: Float?
+    var oils: Float?
+    var milk: Float?
+    var vegetables: Float?
+    var fruits: Float?
 
     
     func getNutrientsAmoumt() {
         switch category {
         case "穀物類", "澱粉類", "飲料類":
-            grains = Double((carbohydrate ?? 0)/15)
+            grains = (carbohydrate ?? 0)/15
             meats = 0.0
             oils = 0.0
             milk = 0.0
@@ -62,7 +62,7 @@ class Food: Mappable {
         case "堅果及種子類", "油脂類":
             grains = 0.0
             meats = 0.0
-            oils = Double((fat ?? 0)/5)
+            oils = (fat ?? 0)/5
             milk = 0.0
             vegetables = 0.0
             fruits = 0.0
@@ -72,17 +72,17 @@ class Food: Mappable {
             oils = 0.0
             milk = 0.0
             vegetables = 0.0
-            fruits = Double((carbohydrate ?? 0)/15)
+            fruits = (carbohydrate ?? 0)/15
         case "蔬菜類", "藻類", "菇類":
             grains = 0.0
             meats = 0.0
             oils = 0.0
             milk = 0.0
-            vegetables = Double((carbohydrate ?? 0)/5)
+            vegetables = (carbohydrate ?? 0)/5
             fruits = 0.0
         case "豆類", "肉類", "魚貝類", "蛋類":
             grains = 0.0
-            meats = Double((protein ?? 0)/7)
+            meats = (protein ?? 0)/7
             oils = 0.0
             milk = 0.0
             vegetables = 0.0
@@ -91,20 +91,20 @@ class Food: Mappable {
             grains = 0.0
             meats = 0.0
             oils = 0.0
-            milk = Double((protein ?? 0)/8)
+            milk = (protein ?? 0)/8
             vegetables = 0.0
             fruits = 0.0
         case "糕餅點心類":
-            grains = Double((carbohydrate ?? 0)/15)
+            grains = (carbohydrate ?? 0)/15
             meats = 0.0
-            oils = Double((fat ?? 0)/5)
+            oils = (fat ?? 0)/5
             milk = 0.0
             vegetables = 0.0
             fruits = 0.0
         case "加工調理食品類":
-            grains = Double((carbohydrate ?? 0)/15)
-            meats = Double((protein ?? 0)/7)
-            oils = Double((fat ?? 0)/5)
+            grains = (carbohydrate ?? 0)/15
+            meats = (protein ?? 0)/7
+            oils = (fat ?? 0)/5
             milk = 0.0
             vegetables = 0.0
             fruits = 0.0
